maya-ai/requirements.txt
fastapi
uvicorn
requests
python-dotenv
jinja2
pydantic[email]
supabase
tweepy
python-multipart
